collector/requirements.txt
SQLAlchemy
psycopg2-binary
stomp.py
orjson